--- src/Components/Router.js
@@ -1,29 +1,0 @@
-import PropTypes from "prop-types";
-import React from "react";
-import { HashRouter as Router, Route, Switch } from "react-router-dom";
-import Feed from "../Routes/Feed";
-import Auth from "../Routes/Auth";
-
-const LoggedInRoutes = () => (
-  <>
-    <Route exact path="/" component={Feed} />
-  </>
-);
-
-const LoggedOutRoutes = () => (
-  <>
-    <Route exact path="/" component={Auth} />
-  </>
-);
-
-const AppRouter = ({ isLoggedIn }) => (
-  <Router>
-    <Switch>{isLoggedIn ? <LoggedInRoutes /> : <LoggedOutRoutes />}</Switch>
-  </Router>
-);
-
-AppRouter.propTypes = {
-  isLoggedIn: PropTypes.bool.isRequired
-};
-
-export default AppRouter;
